+++ models/post.js
@@ -54,13 +54,13 @@ const PostSchema            = new mongoose.Schema({
 // ==============================
 // create and associate (model)
 PostSchema.static('postsCreateThenAssociate', function (raw, user, next) {
-    return Promise.resolve().then(() => _fn.schema.updateAndBind(raw, user, next, 'posts', '$push', this));
+    return _fn.schema.updateAndBind(raw, user, next, 'posts', '$push', this);
 });
 
 
 // delete and dissociate (model)  // note: not workable for admin in deleting media owned by multiple users
 PostSchema.static('postsDeleteThenDissociate', function (docsID, user, next) {
-    return Promise.resolve().then(() => _fn.schema.updateAndBind(docsID, user, next, 'posts', '$pullAll', this));
+    return _fn.schema.updateAndBind(docsID, user, next, 'posts', '$pullAll', this);
 });